+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [FEATURE] Start RUM session on RUM init. See [#1594][]
+
 # 2.5.0 / 08-11-2023
 
 - [BUGFIX] Optimize Session Replay diffing algorithm. See [#1524][]
@@ -557,6 +559,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1524]: https://github.com/DataDog/dd-sdk-ios/pull/1524
 [#1529]: https://github.com/DataDog/dd-sdk-ios/pull/1529
 [#1533]: https://github.com/DataDog/dd-sdk-ios/pull/1533
+[#1594]: https://github.com/DataDog/dd-sdk-ios/pull/1594
 [#1536]: https://github.com/DataDog/dd-sdk-ios/pull/1536
 [#1531]: https://github.com/DataDog/dd-sdk-ios/pull/1531
 [@00fa9a]: https://github.com/00FA9A